PapaBench: bugfix of a stupid bug
...gcc 4.9.1 takes anything without a warning...

+++ benchmarks/PapaBench-0.2m/sw/airborne/autopilot_simulate/mainloop.c
@@ -40,7 +40,7 @@
 // SIM: done
 int main( void )
 {
-  uint8_t l;
+  uint16_t l;
   uint8_t init_cpt;
   /* init peripherals */
   timer_init();
